- Turn off display_errors and log instead; save() no longer returns raw mysqli error text - Validate installer logo filename in get_dominant_color (basename + allowlist) - Restore EMONCMS_EXEC guard in shared rememberme_model.php - admin_switch_user: session_regenerate_id(true) so the old session is destroyed - Reflected XSS: readkey must be 32 hex chars, json_encode at the script sinks - max/min_cap_test load: require has_read_access, not just system existence

+++ www/Modules/system/system_controller.php
@@ -71,11 +71,11 @@ function system_controller() {
     
     if ($route->action=="log" && $session['admin']) {
         if ($route->format=="json") {
-            $system_id = get("id",false);
+            $system_id = (int) get("id",false);
             return $system->get_changes($system_id);
         } else {
             $route->format = "html";
-            $system_id =  get("id",false);
+            $system_id = (int) get("id",false);
             return view("Modules/system/system_log_view.php", array(
                 "system_id"=>$system_id
             ));
